pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "snowflake-connector-python>=3.16.0",
    "streamlit>=1.46.1",
]
